+++ packages/sources/amberdata/src/endpoint/token.ts
@@ -8,8 +8,17 @@ const customError = (data: any) => {
 }
 
 export const inputParameters: InputParameters = {
-  base: ['base', 'from', 'coin'],
-  resultPath: false,
+  base: {
+    required: true,
+    aliases: ['from', 'coin'],
+    description: 'The symbol of the currency to query',
+    type: 'string',
+  },
+  resultPath: {
+    required: false,
+    description: 'The object path to access the value that will be returned as the result',
+    type: 'string',
+  },
 }
 
 export const execute: ExecuteWithConfig<Config> = async (input, _, config) => {