+++ packages/sources/binance-dex/test/integration/adapter.test.ts
@@ -72,7 +72,7 @@ describe('execute', () => {
         .set('Accept', '*/*')
         .set('Content-Type', 'application/json')
         .expect('Content-Type', /json/)
-        .expect(400)
+        .expect(200)
       expect(response.body).toMatchSnapshot()
     })
   })