+++ packages/sources/dns-query/src/endpoint/index.ts
@@ -1,0 +1,1 @@
+export * as dnsQuery from './dnsQuery'