+++ packages/sources/ethgasstation/src/endpoint/gasprice.ts
@@ -4,7 +4,13 @@ import { ExecuteWithConfig, Config, InputParameters } from '@chainlink/types'
 export const supportedEndpoints = ['gasprice']
 
 export const inputParameters: InputParameters = {
-  speed: false,
+  speed: {
+    required: false,
+    description: 'The desired speed',
+    type: 'string',
+    options: ['safeLow', 'average', 'fast', 'fastest'],
+    default: 'average',
+  },
 }
 
 export interface ResponseSchema {
@@ -27,7 +33,7 @@ export const execute: ExecuteWithConfig<Config> = async (request, _, config) => 
   if (validator.error) throw validator.error
 
   const jobRunID = validator.validated.id
-  const speed = validator.validated.data.speed || 'average'
+  const speed = validator.validated.data.speed
   const url = `/api/v1/egs/api/ethgasAPI.json?`
 
   const options = {