+++ packages/sources/ethgaswatch/src/endpoint/gasprice.ts
@@ -4,7 +4,13 @@ import { Config, ExecuteWithConfig, InputParameters } from '@chainlink/types'
 export const supportedEndpoints = ['gasprice']
 
 export const inputParameters: InputParameters = {
-  speed: false,
+  speed: {
+    required: false,
+    description: 'The desired speed',
+    type: 'string',
+    options: ['slow', 'normal', 'fast', 'instant'],
+    default: 'fast',
+  },
 }
 
 export interface ResponseSchema {
@@ -29,7 +35,7 @@ export const execute: ExecuteWithConfig<Config> = async (request, _, config) => 
   if (validator.error) throw validator.error
 
   const jobRunID = validator.validated.id
-  const speed = validator.validated.data.speed || 'fast'
+  const speed = validator.validated.data.speed
   const url = `/api/gas`
 
   const options = {