+++ packages/sources/lition/src/endpoint/energy.ts
@@ -4,9 +4,24 @@ import { ExecuteWithConfig, Config, InputParameters } from '@chainlink/types'
 export const supportedEndpoints = ['energy']
 
 export const inputParameters: InputParameters = {
-  source: true,
-  date: false,
-  hour: false,
+  source: {
+    required: true,
+    options: ['1', '2', '3'],
+    type: 'string',
+    description:
+      'The provider to retrieve price data from (1 - `agora-energiewende.de`, 2 - `smard.de`, 3 - `energy-charts.info`)',
+  },
+  date: {
+    required: false,
+    type: 'string',
+    description:
+      'The date to query formatted by `YYYY-MM-DD` (e.g. `2020-10-12`), defaults to current UTC date',
+  },
+  hour: {
+    required: false,
+    description: 'The hour to query (`0` to `23`), defaults to current UTC hour',
+    type: 'number',
+  },
 }
 
 export interface ResponseSchema {