+++ packages/sources/onchain/src/index.ts
@@ -1,5 +1,12 @@
 import { expose } from '@chainlink/ea-bootstrap'
-import { makeExecute } from './adapter'
+import { makeExecute, endpointSelector } from './adapter'
 import { makeConfig, NAME } from './config'
+import * as types from './endpoint'
 
-export = { NAME, makeExecute, makeConfig, ...expose(NAME, makeExecute()) }
+export = {
+  NAME,
+  makeExecute,
+  makeConfig,
+  ...expose(NAME, makeExecute(), undefined, endpointSelector),
+  types,
+}