+++ packages/sources/wbtc-address-set/src/endpoint/addresses.ts
@@ -1,5 +1,5 @@
 import { Requester, Validator } from '@chainlink/ea-bootstrap'
-import { ExecuteWithConfig } from '@chainlink/types'
+import { ExecuteWithConfig, InputParameters } from '@chainlink/types'
 import { Config } from '../config'
 
 export const supportedEndpoints = ['addresses']
@@ -18,6 +18,7 @@ type Address = {
 }
 
 type AddressType = 'custodial' | 'merchant' | 'deposit'
+export const inputParameters: InputParameters = {}
 
 export const execute: ExecuteWithConfig<Config> = async (request, _, config) => {
   const validator = new Validator(request)