+++ packages/composites/bitcoin-json-rpc/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @chainlink/bitcoin-json-rpc-adapter
 
+## 1.0.16
+
+### Patch Changes
+
+- Updated dependencies [f272a595]
+- Updated dependencies [1d55bbde]
+  - @chainlink/ea-bootstrap@1.6.0
+  - @chainlink/ea-test-helpers@1.0.1
+  - @chainlink/json-rpc-adapter@1.1.11
+
 ## 1.0.15
 
 ### Patch Changes