+++ packages/examples/source/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @chainlink/example-source-adapter
 
+## 1.0.15
+
+### Patch Changes
+
+- Updated dependencies [f272a595]
+- Updated dependencies [1d55bbde]
+  - @chainlink/ea-bootstrap@1.6.0
+  - @chainlink/ea-test-helpers@1.0.1
+
 ## 1.0.14
 
 ### Patch Changes