+++ packages/sources/coinapi/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @chainlink/coinapi-adapter
 
+## 1.1.0
+
+### Minor Changes
+
+- 6b7f9d56: input parameters changed
+
+### Patch Changes
+
+- cf81f1b1: added respone types for crypto endpoint
+- Updated dependencies [f272a595]
+- Updated dependencies [1d55bbde]
+  - @chainlink/ea-bootstrap@1.6.0
+  - @chainlink/ea-test-helpers@1.0.1
+
 ## 1.0.14
 
 ### Patch Changes