+++ packages/sources/durin/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @chainlink/durin-adapter
 
+## 1.1.0
+
+### Minor Changes
+
+- 8a789a49: Input param changes
+
+### Patch Changes
+
+- Updated dependencies [f272a595]
+- Updated dependencies [1d55bbde]
+  - @chainlink/ea-bootstrap@1.6.0
+
 ## 1.0.14
 
 ### Patch Changes